# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/activities:
    get:
      tags:
        - Activities
      summary: List Activities
      description: Get a paginated list of Activities with optional filtering and sorting
      operationId: listActivities
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Activities
      summary: Create Activity
      description: Create a new Activity
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Activity:
      type: object
      required:
        - id
        - subject
        - activity_type
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        activity_type:
          type: string
        direction:
          nullable: true
          type: string
        lead_id:
          nullable: true
          type: string
          format: uuid
        opportunity_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        scheduled_at:
          nullable: true
          type: string
          format: date-time
        occurred_at:
          nullable: true
          type: string
          format: date-time
        assignee_id:
          nullable: true
          type: string
          format: uuid
        outcome:
          nullable: true
          type: string
          maxLength: 280
        notes:
          nullable: true
          type: string
          maxLength: 4000
        metadata:
          type: object
...
    CreateActivityRequest:
      type: object
      required:
        - subject
        - activity_type
        - status
        - metadata
      properties:
        subject:
          type: string
          maxLength: 200
        activity_type:
          type: string
        direction:
          nullable: true
          type: string
        lead_id:
          nullable: true
          type: string
          format: uuid
        opportunity_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        scheduled_at:
          nullable: true
          type: string
          format: date-time
        occurred_at:
          nullable: true
          type: string
          format: date-time
        assignee_id:
          nullable: true
          type: string
          format: uuid
        outcome:
          nullable: true
          type: string
          maxLength: 280
        notes:
          nullable: true
          type: string
          maxLength: 4000
        metadata:
          type: object
...
    ActivityList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []